Create HNG 02113.js
I forked the original repository, then made a clone of the repository so as to add a file without making changes to the original repository. Then i progressed to create a new file under the script directory. it was ensured that the this file was committed to the submission branch, after which i created a pull request.